voice-agent-backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Azure OpenAI Integration
openai==1.3.5

# LangChain for NL to SQL with Azure OpenAI
langchain==0.0.340
langchain-openai==0.0.2
langchain-experimental==0.0.45

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.0.0

# Audio Processing
pydub==0.25.1

# Authentication & Security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
